OS X compile fix after [26255]
git-svn-id: http://svn.mythtv.org/svn/trunk@26260 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/audiooutputca.cpp
@@ -382,7 +382,7 @@ int AudioOutputCA::GetBufferedOnSoundcard(void) const
 /** Reimplement the base class's version of GetAudiotime()
  *  so that we don't use gettimeofday or Qt mutexes.
  */
-int AudioOutputCA::GetAudiotime(void)
+int64_t AudioOutputCA::GetAudiotime(void)
 {
     int audbuf_timecode = GetBaseAudBufTimeCode();
